split v2-ix
