--- anchor_tests/int/v2-ix/Test.toml
@@ -1,4 +1,0 @@
-extends = ["../../setup.toml"]
-
-[scripts]
-test = "yarn vitest run --test-timeout 1000000 --no-file-parallelism --globals legacy-sdk/whirlpool/tests/integration/v2-ix"
