+++ legacy-sdk/whirlpool/tests/integration/v2-ix/collect/collect_reward_v2.test.ts
@@ -3,7 +3,7 @@ import { BN } from "@coral-xyz/anchor";
 import { MathUtil } from "@orca-so/common-sdk";
 import * as assert from "assert";
 import Decimal from "decimal.js";
-import type { WhirlpoolData } from "../../../src";
+import type { WhirlpoolData } from "../../../../src";
 import {
   buildWhirlpoolClient,
   collectRewardsQuote,
@@ -12,8 +12,8 @@ import {
   toTx,
   WhirlpoolContext,
   WhirlpoolIx,
-} from "../../../src";
-import { IGNORE_CACHE } from "../../../src/network/public/fetcher";
+} from "../../../../src";
+import { IGNORE_CACHE } from "../../../../src/network/public/fetcher";
 import {
   approveToken,
   getTokenBalance,
@@ -23,14 +23,14 @@ import {
   TickSpacing,
   transferToken,
   ZERO_BN,
-} from "../../utils";
-import { defaultConfirmOptions } from "../../utils/const";
-import { WhirlpoolTestFixtureV2 } from "../../utils/v2/fixture-v2";
-import type { TokenTrait } from "../../utils/v2/init-utils-v2";
-import { createTokenAccountV2, createMintV2 } from "../../utils/v2/token-2022";
-import { createTokenAccount as createTokenAccountForPosition } from "../../utils/token";
+} from "../../../utils";
+import { defaultConfirmOptions } from "../../../utils/const";
+import { WhirlpoolTestFixtureV2 } from "../../../utils/v2/fixture-v2";
+import type { TokenTrait } from "../../../utils/v2/init-utils-v2";
+import { createTokenAccountV2, createMintV2 } from "../../../utils/v2/token-2022";
+import { createTokenAccount as createTokenAccountForPosition } from "../../../utils/token";
 import { NATIVE_MINT } from "@solana/spl-token";
-import { TokenExtensionUtil } from "../../../src/utils/public/token-extension-util";
+import { TokenExtensionUtil } from "../../../../src/utils/public/token-extension-util";
 
 describe("collect_reward_v2", () => {
   const provider = anchor.AnchorProvider.local(
